refactor: build error fix
